HLE: Remove BaseInterface and add ns.cpp back
