+++ src/core/CMakeLists.txt
@@ -205,6 +205,8 @@ add_library(core STATIC
     hle/service/apt/apt_s.h
     hle/service/apt/apt_u.cpp
     hle/service/apt/apt_u.h
+    hle/service/apt/ns.cpp
+    hle/service/apt/ns.h
     hle/service/apt/ns_s.cpp
     hle/service/apt/ns_s.h
     hle/service/apt/bcfnt/bcfnt.cpp
